+++ DDDEastAnglia/Models/ExternalLogin.cs
@@ -1,0 +1,9 @@
+﻿namespace DDDEastAnglia.Models
+{
+    public class ExternalLogin
+    {
+        public string Provider { get; set; }
+        public string ProviderDisplayName { get; set; }
+        public string ProviderUserId { get; set; }
+    }
+}
